Move response buffering to the transport
